optimize(klisp): Optimize runtiem perf of variable name lookup
Previously, Klisp looked up variables in scope by simply performing a
lookup in a composite (map) and when a () (null) value was found, it
iterated through the keys of the map representing the local environment
to check whether the value was () because it was unbound, or because it
was set to a null value.

The new environment abstraction provides a getenv() and a setenv() that
encapsulates a new implementation detail where Null values are
represented by a runtime-unique 64-bit sentinel value, so no more
iterating through keys of a map is necessary.

This dramatically improves the performance of idiomatic Klisp code on
this interpreter:

klisp test/eval.klisp goes from
        ~3.25s avg to
        ~2.40s avg