fix(nightvale): Fix bug when typing ( on range selection in code
The embedded code editor used to have an incorrect behavior for the '('
keystroke, where if a range selection was active in the input field,
that selection would not be first deleted before () was inserted, as
expected.

This has been fixed. Now typing '(' on top of a range selection removes
the selection and replaces it.